[Release] README, Monitoring Agent, DDL, Queries

+++ linux_sql/sql/ddl.sql
@@ -6,11 +6,11 @@ CREATE TABLE IF NOT EXISTS PUBLIC.host_info
 (
 	id               	SERIAL NOT NULL,
 	hostname         	VARCHAR(100) NOT NULL,
-	cpu_number			SMALLINT NOT NULL,
+	cpu_number			INTEGER NOT NULL,
 	cpu_architecture	VARCHAR(100) NOT NULL,
 	cpu_model			VARCHAR(100) NOT NULL,
 	cpu_mhz				REAL NOT NULL,
-	L2_cache			SMALLINT NOT NULL,
+	L2_cache			INTEGER NOT NULL,
 	total_mem			INTEGER NOT NULL,
 	"timestamp"			TIMESTAMP NOT NULL,
 
@@ -24,10 +24,10 @@ CREATE TABLE IF NOT EXISTS PUBLIC.host_usage
 (
 	"timestamp"    	TIMESTAMP NOT NULL,
 	host_id        	SERIAL NOT NULL,
-	memory_free		SMALLINT NOT NULL,
-	cpu_idle		SMALLINT NOT NULL,
-	cpu_kernel		SMALLINT NOT NULL,
-	disk_io			SMALLINT NOT NULL,
+	memory_free		INTEGER NOT NULL,
+	cpu_idle		INTEGER NOT NULL,
+	cpu_kernel		INTEGER NOT NULL,
+	disk_io			INTEGER NOT NULL,
 	disk_available	INTEGER NOT NULL,
 
 	-- add foreign key constraint
